rename mapper and dependencies update, remove old
